added scrolltotop to searchpage
