+++ src/pages/HomePage/HomePage.tsx
@@ -1,4 +1,4 @@
-import React, { FC, useEffect } from 'react';
+import React, { FC, useEffect, useState } from 'react';
 import { Hero, About, ScrollToTop } from 'components';
 import { useSearchParams } from 'react-router-dom';
 import { updateAccessToken, updateLoginGoogle } from 'redux/auth/authSlice';
@@ -7,6 +7,22 @@ import { useAppDispatch } from 'redux/reduxHooks';
 const HomePage: FC = () => {
   const dispatch = useAppDispatch();
   const [searchParams] = useSearchParams();
+  const [isVisible, setIsVisible] = useState(false);
+
+  const toggleVisibility = () => {
+    if (window.scrollY > 500) {
+      setIsVisible(true);
+    } else {
+      setIsVisible(false);
+    }
+  };
+
+  useEffect(() => {
+    window.addEventListener('scroll', toggleVisibility);
+    return () => {
+      window.removeEventListener('scroll', toggleVisibility);
+    };
+  }, []);
 
   useEffect(() => {
     const accessToken = searchParams.get('accessToken');
@@ -26,7 +42,7 @@ const HomePage: FC = () => {
     <>
       <Hero />
       <About />
-      <ScrollToTop />
+      <ScrollToTop scrollVisible={isVisible} />
     </>
   );
 };
