+++ src/pages/SearchPage/SearchPage.tsx
@@ -1,6 +1,6 @@
-import React, { FC, useState } from 'react';
+import React, { FC, useState, useEffect } from 'react';
 import { useSearchParams } from 'react-router-dom';
-import { SearchInput, SearchList } from 'components';
+import { SearchInput, SearchList, ScrollToTop } from 'components';
 import { Section } from './SearchPage.styled';
 
 const SearchPage: FC = () => {
@@ -8,6 +8,22 @@ const SearchPage: FC = () => {
   const [page, setPage] = useState<number>(1);
   const [query, setQuery] = useState('');
   const [paramsP, setParamsP] = useState(true);
+  const [isVisible, setIsVisible] = useState(false);
+
+  const toggleVisibility = () => {
+    if (window.scrollY > 800) {
+      setIsVisible(true);
+    } else {
+      setIsVisible(false);
+    }
+  };
+
+  useEffect(() => {
+    window.addEventListener('scroll', toggleVisibility);
+    return () => {
+      window.removeEventListener('scroll', toggleVisibility);
+    };
+  }, []);
 
   const paramsQuery: string = searchParams.get('query') ?? `${query}`;
   const paramsSearchType: string =
@@ -26,19 +42,22 @@ const SearchPage: FC = () => {
   };
 
   return (
-    <Section>
-      <SearchInput
-        onSubmit={onSubmit}
-        paramsQuery={paramsQuery}
-        paramsSearchType={paramsSearchType}
-      />
-      <SearchList
-        query={paramsQuery}
-        page={page}
-        loadMore={loadMore}
-        isSearchUsers={paramsSearchType}
-      />
-    </Section>
+    <>
+      <Section>
+        <SearchInput
+          onSubmit={onSubmit}
+          paramsQuery={paramsQuery}
+          paramsSearchType={paramsSearchType}
+        />
+        <SearchList
+          query={paramsQuery}
+          page={page}
+          loadMore={loadMore}
+          isSearchUsers={paramsSearchType}
+        />
+      </Section>
+      <ScrollToTop scrollVisible={isVisible} />
+    </>
   );
 };
